Improves Makefile.
Adds extra backup if the user specifies BUILD_DIR to be empty.
Previously this would have meant a clean would run 'rm -f /*' which is
clearly not a good idea. Now the extra './' before each BUILD_DIR at
least prevents this, and the specified file extensions also help ensure
only temp files are removed.

Also adds an explicity OUTPUT variable to specify the final pdf name.
This can be set in the Makefile, or on the command line to provide
different outputs or just to rename the final pdf.